Fixes: NB#280030 - Navigation bar disappears if page transition happens with orientation change.
RevBy: Stanislav Ionascu
MergeRequest: 2205

Details: Due to changes in latest qt version, property animation
startValue must have the same type as endValue, otherwise animation
doesn't work.

This commit adds casting 0 to qreal in setting start/end values of opacity
animation.

+++ src/corelib/animation/widget/mwidgetfadeanimation.cpp
@@ -105,11 +105,11 @@ void MWidgetFadeAnimation::updateState(QAbstractAnimation::State newState, QAbst
 
         if (d->direction == In) {
             targetWidget()->setOpacity(0);
-            d->opacityAnimation->setStartValue(0);
+            d->opacityAnimation->setStartValue((qreal)0);
             d->opacityAnimation->setEndValue(d->originalOpacity);
         } else {
             d->opacityAnimation->setStartValue(d->originalOpacity);
-            d->opacityAnimation->setEndValue(0);
+            d->opacityAnimation->setEndValue((qreal)0);
         }
 
         d->played = true;
